automation/requirements.txt
# Core dependencies
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
redis==4.5.4
aiohttp==3.8.4
websockets==11.0.3
prometheus-client==0.17.0
psutil==5.9.5
pyotp==2.8.0
qrcode==7.4.2
jinja2==3.1.2
python-dotenv==1.0.0
cryptography==40.0.2
requests==2.31.0
aiofiles==23.1.0
python-socketio==5.8.0
eventlet==0.33.3
gunicorn==21.2.0
supervisor==4.2.5

# Data processing
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
torch==2.0.1
transformers==4.29.2
tensorflow==2.12.0

# Monitoring and logging
grafana-api-client==1.0.3
python-json-logger==2.0.7
flask==2.3.2
plotly==5.14.1

# Documentation
sphinx==6.1.3
sphinx-rtd-theme==1.2.0
mkdocs==1.4.3
mkdocs-material==9.1.3

# Security
python-jose==3.3.0
bcrypt==4.0.1

# Task scheduling
celery==5.3.1
flower==2.0.1

# Model management
mlflow==2.5.0
tensorboard==2.12.3

# Data validation
great-expectations==0.16.13
pandera==0.10.0

# Feature engineering
ta-lib==0.4.28
statsmodels==0.14.0
arch==6.1.0

# Visualization
plotly==5.14.1
dash==2.9.3
dash-bootstrap-components==1.4.1

# API documentation
swagger-ui-bundle==0.0.9
redoc==2.0.0

# Error handling
sentry-sdk==1.25.1
python-json-logger==2.0.7

# Performance optimization
numba==0.57.0
dask==2023.3.2
distributed==2023.3.2

# Database
sqlalchemy==2.0.12
alembic==1.11.1
asyncpg==0.27.0
psycopg2-binary==2.9.6

# Caching
redis-py-cluster==2.1.3
memcached==1.0.0

# Message queue
kafka-python==2.0.2
confluent-kafka==2.1.1

# Containerization
docker==6.1.2
kubernetes==26.1.0

# Cloud integration
boto3==1.26.137
google-cloud-storage==2.9.0
azure-storage-blob==12.16.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tenacity==8.2.2
structlog==23.1.0

# Additional dependencies
aiosmtplib==2.0.2 